requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
streamlit-cookies-controller
